removed editing from team page
